Implemented FFDH with rotations.

+++ src/nl/tue/algorithms/dbl/io/OutputWriter.java
@@ -30,7 +30,7 @@ public class OutputWriter {
             } else {
                 systemOut.print("no ");
             }
-            systemOut.println(r.getX()+ " " +r.getY());
+            systemOut.println((int)r.getX()+ " " +(int)r.getY());
         }
     }
 }
